SpecialLottery v2, unittest, deploy script (#11)
* testnet manifest (with standard v2 contract)

* SpecialLottery v2, unittest, deploy script

* fix: missing await and formatting on special-lottery-upgradability.js

* test: stadard lottery - always have to have a test for the potentialy next upgrade

e.x.: if we have V2 we should make a fake V3 and create a upgradeability test from V2 to V3. Because it's easy to make a mistake and break future upgrades.

* test: special lottery - always have to have a test for the potentialy next upgrade

e.x.: if we have V2 we should make a fake V3 and create a upgradeability test from V2 to V3. Because it's easy to make a mistake and break future upgrades.

Co-authored-by: Ben Weider <ben.weider081@yahoo.com>
Co-authored-by: Aleksandr Pasevin <pasevin@gmail.com>